Updated default adv searches to use adv criteria instead of simple so they should be more convieneint and kinda show how adv works. Reset data and began testing removing message data in transactions and it works but thinking transactions are a dead end..the metaphor isn't really carrying through well and just don't know what to do there but now its not dumping extra data all the time, only w/ annotations so text works but images doesn't since its a canvas overlay so gonna back to that I think it'd be better if it uses blots so its part of the html but thats a whole nother headache...Also was brave enough to split out the enumerable parameter types into seperate subclasses instead of jumbling everything together and now using avalonia selectionModel which isn't very good but now they're seperated and responding so that's good.

+++ Common/MonkeyPaste.Common/Utils/Extensions/MpCommonExtensions.cs
@@ -660,7 +660,7 @@ namespace MonkeyPaste.Common {
             }
             if (obj is string strObj) {
                 try {
-                    return int.Parse(strObj);
+                    return (int)double.Parse(strObj);
                 }
                 catch (Exception ex) {
                     MpConsole.WriteTraceLine($"Error parsing int from '{strObj}'", ex);
